third_party/WebKit/LayoutTests/inspector/console/console-uncaught-promise-expected.txt

CONSOLE ERROR: line 19: Uncaught (in promise) Error: err1
CONSOLE ERROR: line 38: Uncaught (in promise) Error: err2
CONSOLE ERROR: line 54: Uncaught (in promise) NotFoundError: Failed to execute 'removeChild' on 'Node': The node to be removed is not a child of this node.
CONSOLE ERROR: line 59: Uncaught (in promise) 42
CONSOLE ERROR: line 63: Uncaught (in promise) 1e+100
CONSOLE ERROR: line 67: Uncaught (in promise) foo
CONSOLE ERROR: line 71: Uncaught (in promise) #<Object>
CONSOLE ERROR: line 75: Uncaught (in promise) undefined
CONSOLE ERROR: line 79: Uncaught (in promise) SecurityError: Failed to register a ServiceWorker: The URL protocol of the current origin ('') is not supported.
Tests that uncaught promise rejections are logged into console.

console-uncaught-promise.html:19 Uncaught (in promise) Error: err1
promiseTest1 @ console-uncaught-promise.html:19
timeout @ console-uncaught-promise.html:13
Promise rejected (async)
promiseTest1 @ console-uncaught-promise.html:19
timeout @ console-uncaught-promise.html:13
setTimeout (async)
runNextPromiseTest @ console-uncaught-promise.html:12
(anonymous) @ VM:1
console-uncaught-promise.html:38 Uncaught (in promise) Error: err2
promiseTest2 @ console-uncaught-promise.html:38
timeout @ console-uncaught-promise.html:13
Promise rejected (async)
(anonymous) @ console-uncaught-promise.html:34
Promise rejected (async)
(anonymous) @ console-uncaught-promise.html:33
Promise rejected (async)
(anonymous) @ console-uncaught-promise.html:32
Promise rejected (async)
promiseTest2 @ console-uncaught-promise.html:31
timeout @ console-uncaught-promise.html:13
setTimeout (async)
runNextPromiseTest @ console-uncaught-promise.html:12
(anonymous) @ VM:1
console-uncaught-promise.html:54 Uncaught (in promise) DOMException: Failed to execute 'removeChild' on 'Node': The node to be removed is not a child of this node.
throwDOMException @ console-uncaught-promise.html:54
catcher @ console-uncaught-promise.html:47
Promise rejected (async)
promiseTest3 @ console-uncaught-promise.html:46
timeout @ console-uncaught-promise.html:13
setTimeout (async)
runNextPromiseTest @ console-uncaught-promise.html:12
(anonymous) @ VM:1
console-uncaught-promise.html:59 Uncaught (in promise) 42
promiseTest4 @ console-uncaught-promise.html:59
timeout @ console-uncaught-promise.html:13
setTimeout (async)
runNextPromiseTest @ console-uncaught-promise.html:12
(anonymous) @ VM:1
console-uncaught-promise.html:63 Uncaught (in promise) 1e+100
promiseTest5 @ console-uncaught-promise.html:63
timeout @ console-uncaught-promise.html:13
setTimeout (async)
runNextPromiseTest @ console-uncaught-promise.html:12
(anonymous) @ VM:1
console-uncaught-promise.html:67 Uncaught (in promise) foo
promiseTest6 @ console-uncaught-promise.html:67
timeout @ console-uncaught-promise.html:13
setTimeout (async)
runNextPromiseTest @ console-uncaught-promise.html:12
(anonymous) @ VM:1
console-uncaught-promise.html:71 Uncaught (in promise) {foo: 42}
    foo: 42
    __proto__: Object
promiseTest7 @ console-uncaught-promise.html:71
timeout @ console-uncaught-promise.html:13
setTimeout (async)
runNextPromiseTest @ console-uncaught-promise.html:12
(anonymous) @ VM:1
console-uncaught-promise.html:75 Uncaught (in promise) undefined
promiseTest8 @ console-uncaught-promise.html:75
timeout @ console-uncaught-promise.html:13
setTimeout (async)
runNextPromiseTest @ console-uncaught-promise.html:12
(anonymous) @ VM:1
console-uncaught-promise.html:79 Uncaught (in promise) DOMException: Failed to register a ServiceWorker: The URL protocol of the current origin ('file://') is not supported.
promiseTest9 @ console-uncaught-promise.html:79
timeout @ console-uncaught-promise.html:13
Promise (async)
promiseTest9 @ console-uncaught-promise.html:79
timeout @ console-uncaught-promise.html:13
setTimeout (async)
runNextPromiseTest @ console-uncaught-promise.html:12
(anonymous) @ VM:1
